chore: adjust error handler to emit internal verify error

+++ contracts/ics/connection/src/main.rs
@@ -11,6 +11,6 @@ default_alloc!();
 fn program_entry() -> i8 {
     match entry::main() {
         Ok(_) => 0,
-        Err(err) => err as i8,
+        Err(err) => err,
     }
 }
